requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-compat==1.0.15
django-rest-enumfield==0.2.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
django-cors-headers==3.8.0